+++ waltid-services/waltid-issuer-api/src/main/kotlin/id/walt/issuer/Main.kt
@@ -5,7 +5,6 @@ import id.walt.commons.ServiceInitialization
 import id.walt.commons.ServiceMain
 import id.walt.commons.featureflag.FeatureManager.whenFeature
 import id.walt.commons.web.WebService
-import id.walt.commons.web.plugins.configureSerialization
 import id.walt.did.helpers.WaltidServices
 import id.walt.issuer.entra.entraIssuance
 import id.walt.issuer.issuance.OidcApi.oidcApi
@@ -32,7 +31,6 @@ fun Application.configurePlugins() {
     configureHTTP()
     configureMonitoring()
     configureRouting()
-    configureSerialization()
 }
 
 fun Application.issuerModule(withPlugins: Boolean = true) {
