+++ waltid-services/waltid-verifier-api/src/main/kotlin/id/walt/verifier/Main.kt
@@ -39,7 +39,6 @@ fun Application.configurePlugins() {
     configureHTTP()
     configureMonitoring()
     configureRouting()
-    configureSerialization()
 }
 
 fun Application.verifierModule(withPlugins: Boolean = true) {
