+++ waltid-services/waltid-verifier-api/src/main/kotlin/id/walt/verifier/lspPotential/LspPotentialVerificationInterop.kt
@@ -10,8 +10,13 @@ import com.nimbusds.jose.jwk.Curve
 import com.nimbusds.jose.jwk.ECKey
 import com.nimbusds.jose.jwk.JWK
 import com.nimbusds.jose.jwk.gen.ECKeyGenerator
+import com.nimbusds.jose.util.X509CertChainUtils
+import id.walt.credentials.vc.vcs.W3CVC
+import id.walt.crypto.keys.KeySerialization
+import id.walt.crypto.keys.jwk.JWKKey
 import id.walt.mdoc.COSECryptoProviderKeyInfo
 import id.walt.mdoc.SimpleCOSECryptoProvider
+import id.walt.mdoc.cose.X5_CHAIN
 import id.walt.mdoc.dataelement.DataElement
 import id.walt.mdoc.dataelement.FullDateElement
 import id.walt.mdoc.dataelement.toDataElement
